+++ x-pack/agent/pkg/fleetapi/action.go
@@ -8,7 +8,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"strings"
-	"time"
 
 	"github.com/pkg/errors"
 )
@@ -34,38 +33,6 @@ func (a *BaseAction) Type() string {
 // ID returns the action ID.
 func (a *BaseAction) ID() string {
 	return a.ActionID
-}
-
-// UnknownAction is an action that is not know by the current version of the Agent and we don't want
-// to return an error at parsing time but at execution time we can report or ignore.
-//
-// NOTE: We only keep the original type and the action id, the payload of the event is dropped, we
-// do this to make sure we do not leak any unwanted information.
-type UnknownAction struct {
-	*BaseAction
-	originalType string
-}
-
-// Type returns the type of the Action.
-func (a *UnknownAction) Type() string {
-	return "UNKNOWN"
-}
-
-func (a *UnknownAction) String() string {
-	var s strings.Builder
-	s.WriteString("action_id: ")
-	s.WriteString(a.ID())
-	s.WriteString(", type: ")
-	s.WriteString(a.Type())
-	s.WriteString(" (original type: ")
-	s.WriteString(a.OriginalType())
-	s.WriteString(")")
-	return s.String()
-}
-
-// OriginalType returns the original type of the action as returned by the API.
-func (a *UnknownAction) OriginalType() string {
-	return a.originalType
 }
 
 // PolicyChangeAction is a request to apply a new
@@ -126,39 +93,4 @@ func (a *Actions) UnmarshalJSON(data []byte) error {
 
 	*a = actions
 	return nil
-}
-
-// AckedAction represents a event to be send to the next checkin that will Ack an action.
-type AckedAction struct {
-	EventType string    `json:"type"`
-	ActionID  string    `json:"action_id"`
-	Ts        time.Time `json:"timestamp"`
-	Msg       string    `json:"message,omitempty"`
-}
-
-// Type return the type of event.
-func (a *AckedAction) Type() string {
-	return a.EventType
-}
-
-// Timestamp return when the event was created.
-func (a *AckedAction) Timestamp() time.Time {
-	return a.Ts
-}
-
-// Message returns the human readable string describing the event.
-func (a *AckedAction) Message() string {
-	return a.Msg
-}
-
-// Ack returns an event that represent an acked action.
-func Ack(action Action) *AckedAction {
-	const t = "ACTION_ACKNOWLEDGED"
-
-	return &AckedAction{
-		EventType: t,
-		ActionID:  action.ID(),
-		Msg:       "Acknowledge action " + action.ID(),
-		Ts:        time.Now(),
-	}
 }
